Move BufferId into its own module.
